какой-то баг с кодировкой

+++ Lab3/main_window.py
@@ -19,6 +19,8 @@ from PyQt6.QtCore import Qt
 from PyQt6.QtGui import QPixmap
 from PyQt6.QtCore import Qt
 
+from Lab2 import year_split_2
+from Lab2 import week_split_3
 from Lab2 import return_4
 
 
@@ -29,11 +31,12 @@ class MainWindow(QMainWindow):
         self.years = None
         self.button_annots = None
         self.button_dir = None
-        # self.path_to_dataset = 'data/dataset.csv'
+
         self.path_to_dataset = ''
-        # self.path_to_annots = 'data/'
         self.path_to_dir = ''
         self.path_to_annots = ''
+        self.min_date = None
+        self.max_date = None
         self.initui()
 
     def initui(self):
@@ -43,6 +46,8 @@ class MainWindow(QMainWindow):
         while self.path_to_dataset == '':
             self.path_to_dataset = QFileDialog.getOpenFileName(self, 'Выберите файл с данными', '',
                                                                'CSV files (*.csv)')[0]
+        self.min_max_date()
+        print(self.min_date, self.max_date)
 
         self.button_annots = QPushButton('Создать файл аннотации', self)
         self.button_annots.move(20, 70)
@@ -62,28 +67,34 @@ class MainWindow(QMainWindow):
         self.button_dir.show()
 
         self.years = QLineEdit('Введите количество лет', self)
-        self.years.move(20, 120)
+        self.years.move(20, 130)
         self.years.resize(200, 20)
         self.years.show()
 
         self.weeks = QLineEdit('Введите количество недель', self)
-        self.weeks.move(20, 150)
+        self.weeks.move(20, 160)
         self.weeks.resize(200, 20)
         self.weeks.show()
-
-    def create_annots(self):
-        # copy path_to_dataset to path_to_annots/dataset.csv.temp
-
-        shutil.copy(self.path_to_dataset, self.path_to_dir + '/dataset.csv.temp')
-        self.path_to_annots = self.path_to_dir + '/dataset.csv.temp'
-
-        self.button_annots.setText('Файл аннотации создан')
 
     def open_dir(self):
         self.path_to_dir = QFileDialog.getExistingDirectory(self, 'Выберите директорию', '')
         self.button_dir.setText('Директория выбрана')
         self.button_dir.setEnabled(False)
         self.button_annots.setEnabled(True)
+
+    def create_annots(self):
+        self.path_to_annots = self.path_to_dir + '/dataset.csv.temp'
+        shutil.copy(self.path_to_dataset, self.path_to_annots)
+        self.button_annots.setText('Файл аннотации создан')
+        self.button_annots.setEnabled(False)
+
+    def min_max_date(self):
+        with open(self.path_to_dataset, 'r') as f:
+            reader = csv.reader(f)
+            rows = list(reader)
+            rows = rows[2:]
+            self.min_date = datetime.strptime(rows[0][0], '%Y-%m-%d')
+            self.max_date = datetime.strptime(rows[-1][0], '%Y-%m-%d')
 
 
 if __name__ == '__main__':
@@ -92,5 +103,7 @@ if __name__ == '__main__':
     window.show()
     app.exec()
 
-    # detele temp file
-    os.remove(window.path_to_annots)
+    print(window.weeks, window.years, window.button_annots, window.button_dir, window.path_to_dataset, window.path_to_dir, window.path_to_annots, window.min_date, window.max_date)
+
+    if window.path_to_annots != '':
+        os.remove(window.path_to_annots)
